src/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
autopep8==2.0.4
beautifulsoup4==4.11.2
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.1
django-dirtyfields==1.9.2
django-filter==23.5
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-silk==5.1.0
django-taggit==4.0.0
django-treebeard==4.7.1
djangorestframework==3.14.0
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filetype==1.2.0
gprof2dot==2022.7.29
html5lib==1.1
idna==3.6
l18n==2021.3
openpyxl==3.1.2
pillow==10.2.0
pillow-heif==0.14.0
pycodestyle==2.11.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.2.0
wagtail==5.2.3
webencodings==0.5.1
Willow==1.6.3
